apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.29
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic-settings==2.2.1
redis[hiredis]==5.0.3
celery[redis]==5.3.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
openai==1.30.0
anthropic==0.26.0
langchain==0.2.0
langchain-openai==0.1.7
langchain-anthropic==0.1.11
langgraph==0.1.1
pgvector==0.2.5
numpy==1.26.4
boto3==1.34.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.128.0
sentry-sdk[fastapi]==2.2.0
structlog==24.1.0
tiktoken==0.7.0
pypdf==4.2.0
python-docx==1.1.2
python-dateutil==2.9.0
reportlab==4.2.2
